group rust files in frontend and othello folders
