+++ src/frontend/mod.rs
@@ -1,1 +1,2 @@
+pub mod app;
 pub mod websocket;
