+++ src/main.rs
@@ -1,27 +1,9 @@
-use axum::{extract::ws::WebSocketUpgrade, response::IntoResponse, routing::get, Router};
-use frontend::websocket::handle_socket;
-use std::net::SocketAddr;
-use tower_http::services::ServeDir;
+use frontend::app::run_app;
 
-pub mod board;
 pub mod frontend;
-pub mod position;
+pub mod othello;
 
 #[tokio::main]
 async fn main() {
-    let app = Router::new()
-        .route("/ws", get(ws_handler))
-        .nest_service("/", ServeDir::new("assets"));
-
-    let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
-    println!("Listening on {}", addr);
-
-    axum::Server::bind(&addr)
-        .serve(app.into_make_service())
-        .await
-        .unwrap();
-}
-
-async fn ws_handler(ws: WebSocketUpgrade) -> impl IntoResponse {
-    ws.on_upgrade(handle_socket)
+    run_app().await;
 }
