+++ src/othello/mod.rs
@@ -1,0 +1,2 @@
+pub mod board;
+pub mod position;
